move data -> ./model
