+++ src/code/dialogs/authDialog.js
@@ -10,7 +10,7 @@ import PromptDialog from "./promptDialog";
 import { sendLocation, fullSync } from "../uiCommands";
 import { wX } from "../wX";
 import { postToFirebase } from "../firebaseSupport";
-import WasabeeMe from "../me";
+import WasabeeMe from "../model/me";
 
 const AuthDialog = WDialog.extend({
   statics: {
