+++ src/code/dialogs/homogeneous.js
@@ -1,8 +1,8 @@
 import { WDialog } from "../leafletClasses";
-import WasabeePortal from "../portal";
+import WasabeePortal from "../model/portal";
 import { getSelectedOperation } from "../selectedOp";
 import { greatCircleArcIntersectByLatLngs } from "../crosslinks";
-// import WasabeeLink from "../link";
+// import WasabeeLink from "../model/link";
 import {
   clearAllLinks,
   getAllPortalsOnScreen,
