+++ src/code/dialogs/importDialog.js
@@ -1,5 +1,5 @@
-import WasabeeOp from "../operation";
-import WasabeePortal from "../portal";
+import WasabeeOp from "../model/operation";
+import WasabeePortal from "../model/portal";
 import { WDialog } from "../leafletClasses";
 import OperationChecklistDialog from "./checklist";
 import wX from "../wX";
