+++ src/code/dialogs/opsDialog.js
@@ -1,4 +1,4 @@
-import WasabeeOp from "../operation";
+import WasabeeOp from "../model/operation";
 import { WDialog } from "../leafletClasses";
 import {
   getSelectedOperation,
@@ -12,8 +12,8 @@ import {
 } from "../selectedOp";
 import OpPermList from "./opPerms";
 import wX from "../wX";
-import WasabeeMe from "../me";
-import WasabeeAgent from "../agent";
+import WasabeeMe from "../model/me";
+import WasabeeAgent from "../model/agent";
 import { syncOp, deleteLocalOp, zoomToOperation } from "../uiCommands";
 import Sortable from "../sortable";
