+++ src/code/dialogs/sendTargetDialog.js
@@ -1,8 +1,8 @@
 import { WDialog } from "../leafletClasses";
-import WasabeeMarker from "../marker";
-import WasabeeAnchor from "../anchor";
-import WasabeeMe from "../me";
-import WasabeeTeam from "../team";
+import WasabeeMarker from "../model/marker";
+import WasabeeAnchor from "../model/anchor";
+import WasabeeMe from "../model/me";
+import WasabeeTeam from "../model/team";
 import { targetPromise } from "../server";
 import wX from "../wX";
 import { getSelectedOperation } from "../selectedOp";
